# api/docs/swagger.yaml
basePath: /api
definitions:
...
  middlewares.Unauthorized:
    properties:
      error:
        type: string
    type: object
...
  models.OcservGroup:
    properties:
      config:
        $ref: '#/definitions/models.OcservGroupConfig'
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
  models.OcservGroupConfig:
    properties:
      cgroup:
        description: 'Linux control group to assign the VPN worker process to. Format:
          ''controller,subsystem:name''. Example: ''cpuset,cpu:test'''
        type: string
      deny-roaming:
        description: 'Disconnect client if its IP changes (e.g., due to network switch).
          Example: true'
        type: boolean
      dns:
        description: 'Comma-separated list of DNS servers to assign to the client.
          Example: ''8.8.8.8,1.1.1.1'''
        items:
          type: string
        type: array
      dpd:
        description: 'Dead Peer Detection timeout in seconds. Example: 90'
        type: integer
      explicit-ipv4:
        description: 'Static IPv4 address to assign to client. Example: ''192.168.100.10'''
        type: string
      idle-timeout:
        description: 'Time in seconds before disconnecting idle clients. Example:
          600'
        type: integer
      ipv4-network:
        description: 'The pool of addresses that leases will be given from. Example:
          ''192.168.1.0/24'''
        type: string
      iroute:
        description: 'Internal route available only via VPN. Format: ''IP/prefix''.
          Example: ''10.0.0.0/8'''
        type: string
      keepalive:
        description: 'Interval in seconds to send keep-alive pings. Example: 60'
        type: integer
      max-same-clients:
        description: 'Maximum simultaneous logins per user. Example: 2'
        type: integer
      mobile-dpd:
        description: 'DPD timeout specifically for mobile clients. Example: 300'
        type: integer
      mobile-idle-timeout:
        description: 'Idle timeout for mobile clients. Example: 900'
        type: integer
      mtu:
        description: 'Tunnel interface MTU to avoid fragmentation. Example: 1400'
        type: integer
      nbns:
        description: 'NetBIOS Name Servers (WINS) for Windows clients. Example: ''192.168.1.1'''
        type: string
      net-priority:
        description: 'Priority for routes; lower is higher priority. Example: 1'
        type: integer
      no-route:
        description: 'List of networks to exclude from VPN routing. Each entry should
          be in ''IP/prefix'' format. Example: [''192.168.0.0/16'', ''10.0.0.0/8'']'
        items:
          type: string
        type: array
      no-udp:
        description: 'Disables UDP, enforcing TCP-only VPN connection. Example: true'
        type: boolean
      restrict-user-to-ports:
        description: 'Comma-separated list of allowed (or blocked, if negated) protocols
          and ports. Supports ''tcp(port)'', ''udp(port)'', ''icmp()'', ''icmpv6()'',
          and negation with ''!()''. Example: ''tcp(443), tcp(80), udp(53)'', or ''!(tcp(22),
          udp(1194))'''
        type: string
      restrict-user-to-routes:
        description: 'Allow client access only to defined routes. Example: true'
        type: boolean
      route:
        description: 'Routes pushed to the client for routing traffic. Example: [''0.0.0.0/0'',
          ''10.10.0.0/16'']'
        items:
          type: string
        type: array
      rx-data-per-sec:
        description: 'Maximum receive bandwidth in bytes per second. Example: ''100000''
          for 100 KB/s'
        type: integer
      session-timeout:
        description: 'Max session time in seconds before forced disconnect. Example:
          3600'
        type: integer
      split-dns:
        description: 'List of domains over which the provided DNS servers should be
          used. Example: [''example.com'', ''internal.company.com'']'
        items:
          type: string
        type: array
      stats-report-time:
        description: 'Interval in seconds for stats reporting. Example: 300'
        type: integer
      tunnel-all-dns:
        description: 'Force all DNS traffic through the VPN tunnel. Example: true'
        type: boolean
      tx-data-per-sec:
        description: 'Maximum transmit bandwidth in bytes per second. Example: ''200000''
          for 200 KB/s'
        type: integer
    type: object
...
  ocserv_group.CreateOcservGroupData:
    properties:
      config:
        $ref: '#/definitions/models.OcservGroupConfig'
      name:
        type: string
    required:
    - config
    - name
    type: object
  ocserv_group.OcservGroupsResponse:
    properties:
      meta:
        $ref: '#/definitions/request.Meta'
      result:
        items:
          $ref: '#/definitions/models.OcservGroup'
        type: array
    required:
    - meta
    type: object
  ocserv_group.UpdateOcservGroupData:
    properties:
      config:
        $ref: '#/definitions/models.OcservGroupConfig'
    required:
    - config
    type: object
...
  request.ErrorResponse:
    properties:
      error:
        items:
          type: string
        type: array
      message:
        items:
          type: string
        type: array
    required:
    - error
    - message
    type: object
  request.Meta:
    properties:
      page:
        type: integer
      size:
        type: integer
      total_records:
        type: integer
    required:
    - page
    - size
    - total_records
    type: object
...
info:
  contact: {}
  description: This is a sample Ocserv User management Api server.
  title: Ocserv User management Example Api
  version: "1.0"
paths:
...
  /ocserv/groups:
    get:
      consumes:
      - application/json
      description: List of Ocserv groups
      parameters:
      - description: Page number, starting from 1
        in: query
        minimum: 1
        name: page
        type: integer
      - description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: size
        type: integer
      - description: Field to order by
        in: query
        name: order
        type: string
      - description: Sort order, either ASC or DESC
        enum:
        - ASC
        - DESC
        in: query
        name: sort
        type: string
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ocserv_group.OcservGroupsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: List of Ocserv groups
      tags:
      - Ocserv(Groups)
    post:
      consumes:
      - application/json
      description: Ocserv Group creation
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: ocserv group create data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_group.CreateOcservGroupData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OcservGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Group creation
      tags:
      - Ocserv(Groups)
  /ocserv/groups/{id}:
    delete:
      consumes:
      - application/json
      description: Ocserv Group delete
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv Group ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Group delete
      tags:
      - Ocserv(Groups)
    patch:
      consumes:
      - application/json
      description: Ocserv Group update
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: Ocserv Group ID
        in: path
        name: id
        required: true
        type: integer
      - description: ocserv group create data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_group.UpdateOcservGroupData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.OcservGroup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Group update
      tags:
      - Ocserv(Groups)
  /ocserv/groups/defaults:
    get:
      consumes:
      - application/json
      description: Ocserv Defaults Group config
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Ocserv Defaults Group config
      tags:
      - Ocserv(Groups)
    patch:
      consumes:
      - application/json
      description: Update Ocserv Defaults Group
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      - description: ocserv group default data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ocserv_group.UpdateOcservGroupData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: Update Ocserv Defaults Group
      tags:
      - Ocserv(Groups)
  /ocserv/groups/lookup:
    get:
      consumes:
      - application/json
      description: List of Ocserv group names
      parameters:
      - description: Bearer TOKEN
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/request.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.Unauthorized'
      summary: List of Ocserv group names
      tags:
      - Ocserv(Groups)
...
swagger: "2.0"