# api/internal/api/ocserv_group/types.go
package ocserv_group

import (
	"api/internal/models"
	"api/pkg/request"
)

type CreateOcservGroupData struct {
	Name   string                    `json:"name" validate:"required"`
	Config *models.OcservGroupConfig `json:"config" validate:"required"`
}

type UpdateOcservGroupData struct {
	Config *models.OcservGroupConfig `json:"config" validate:"required"`
}

type OcservGroupsResponse struct {
	Meta   request.Meta          `json:"meta" validate:"required"`
	Result *[]models.OcservGroup `json:"result" validate:"omitempty"`
}